BACKEND/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.25.2
openai>=1.0.0
google-generativeai>=0.8.0
